+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This projec
 - Added the following helper functions:
   - `json_decode()`: wraps the native function and throws a JSON API exception if decoding fails.
   - `http_contains_body()`: determines if a PSR message contains body content.
+- New serializer interface to type-hint that an encoder can also serialize data to an array.
 
 ### Fixed
 - A decoder is now only matched if the client has sent a HTTP request with message body.